SSG-3071: JESM: The value of the throughput metric is not schema valid. Throughput is now a integer value.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@23769 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/gateway/server/src/main/java/com/l7tech/server/wsdm/MetricsRequestContext.java
@@ -24,7 +24,7 @@ public class MetricsRequestContext {
     private final boolean operational;
     private final long serviceId;
     private final long serviceTime;
-    private final double throughput;
+    private final long throughput;
     private final long periodStart;
     private final long lastUpdated;
     private final String duration;
@@ -75,7 +75,7 @@ public class MetricsRequestContext {
         }
 
         this.throuputPeriod = tperiod;
-        this.throughput = tval;
+        this.throughput = (long)tval;
     }
 
     public long getAvgResponseTime() {
@@ -114,7 +114,7 @@ public class MetricsRequestContext {
         return serviceTime;
     }
 
-    public double getThroughput() {
+    public long getThroughput() {
         return throughput;
     }
